move Fragments to respective packages
