+++ sampleapp/src/main/res/layout/fragment_settings_tor.xml
@@ -72,6 +72,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="io.matthewnelson.sampleapp.ui.fragments.settings.SettingsTorFragment">
+    tools:context="io.matthewnelson.sampleapp.ui.fragments.settings.tor.SettingsTorFragment">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
